requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-groq
langchain-chroma
chromadb
streamlit
pandas
python-dotenv
sentence-transformers